[framework] add the capability for creating a signer for a resource account
I was going to add both the signer and move to, but the move to doesn't
work due to Move requiring that resources are defined in the module in
which you use global storage :/. Probably need to explore a native
function for this.

A resource account is used to manage resources independent of an account
managed by a user.  Some applications might include managing a Dao and
the resources associated with the Dao or a liquidity pool to create new
liquidity pool coins without requiring the original account owner to
setup the resources. While one could theoretically manage a lot of this
via capabilities, the `move_to` semantics insist on a signer and cannot
work with capabilities due to Move requirements that `move_to` be
executed on a resource within the same module that defines that
resource.

As a small example:
```
let (signer, cap) = create_resource_account(&source);
let lp = LiquidityPool { signer_cap: cap, ... };
move_to(&signer, lp);
```

Later on during a coin registration:
```
public fun add_coin<X, Y>(lp: &LP, x: Coin<x>, y: Coin<y>) {
   if(!exists<LiquidityCoin<X, Y>(LP::Address(lp), LiquidityCoin<X, Y>)) {
       let mint, burn = Coin::initialize<LiquidityCoin<X, Y>>(...);
       move_to(&create_signer_with_capability(&lp.cap), LiquidityCoin<X, Y>{ mint, burn });
   }
   ...
}
```